+++ OwlOSC/Test/Test.cs
@@ -1,0 +1,13 @@
+
+using System;
+
+namespace OwlOSC{
+    public class Test {
+
+        void Main(){
+            OscPacket packet = new OscMessage("/",1);
+            if(packet.IsBundle){}
+        }
+        
+    }
+}