+++ CHANGELOG.md
@@ -2,5 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1] 2022.08.16
+
+ - Added test console program in `OwlOSC.Test` (send / receive, receive loop Async, speedtest, send / receive file)
+ - Message / Bundle proper 'ToString' override
+ - UDP send size limit (64k)
+ - Simple Bundle recognition with boolean in OSCPacket
+
 ## [1.0] 2022.08.14
+
  - Base OSC from SharpOSC and rename namespace